requirements.txt
alembic==1.10.2
anyio==3.6.2
attrs==22.2.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.0
fastapi==0.95.0
fastapi-utils==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
load-dotenv==0.1.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
protobuf==3.20.3
pydantic==1.10.7
PyMySQL==1.0.2
pytest==7.2.2
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
